Fixed small bug and hooked testing into standard

+++ aruw-mcb-project/src/aruwsrc/control/chassis/fixed_path_auto_nav_command.cpp
@@ -43,7 +43,7 @@ FixedPathAutoNavCommand::FixedPathAutoNavCommand(
       autoNavController(autoNavController),
       path()
 {
-    for (int i = 0; i < sizeof(pathPoints); i++)
+    for (size_t i = 0; i < sizeof(pathPoints); i++)
     {
         path.pushPoint(pathPoints[i]);
     }
